Handling exception in embed externals and open images even if the file already exists.

+++ myModules.py
@@ -279,22 +279,25 @@ def dump_html(
             my_embed_external_path_relative = os.path.join(str('../' + my_vars['attach_dir']),my_embed_external_name)
         else:
             my_embed_external_path_relative = os.path.join(my_vars['attach_dir'],my_embed_external_name)
-        to_download = requests.get(orig_embed_external_path, allow_redirects=True)
         try:
-            open(my_embed_external_path,'wb').write(to_download.content)
+            if not os.path.exists(my_embed_external_path):
+                to_download = requests.get(orig_embed_external_path, allow_redirects=True)
+                open(my_embed_external_path,'wb').write(to_download.content)
+            img = Image.open(my_embed_external_path)
         except:
-            print(orig_embed_external_path)
-        img = Image.open(my_embed_external_path)
-        if img.width < 600:
-            embed_ext['width'] = img.width
+            print(f"WARNING: Skipping embed file {my_embed_external_path} due to issues. url: {orig_embed_external_path}")
         else:
-            embed_ext['width'] = 600
-        img.close
-        embed_ext['height'] = "auto"
-        embed_ext['onclick'] = f"window.open(\"{my_embed_external_path_relative}\")"
-        embed_ext['src'] = str(my_embed_external_path_relative)
-        embed_ext['data-image-src'] = str(my_embed_external_path_relative)
-        my_embeds_externals_counter = my_embeds_externals_counter + 1
+            if img is not None:
+                if img.width < 600:
+                    embed_ext['width'] = img.width
+                else:
+                    embed_ext['width'] = 600
+                img.close
+                embed_ext['height'] = "auto"
+                embed_ext['onclick'] = f"window.open(\"{my_embed_external_path_relative}\")"
+                embed_ext['src'] = str(my_embed_external_path_relative)
+                embed_ext['data-image-src'] = str(my_embed_external_path_relative)
+                my_embeds_externals_counter = my_embeds_externals_counter + 1
 
     #
     # dealing with "confluence-embedded-image"
@@ -315,7 +318,7 @@ def dump_html(
             if not os.path.exists(my_embed_path):
                 to_download = requests.get(orig_embed_path, allow_redirects=True, auth=(arg_username, arg_api_token))
                 open(my_embed_path,'wb').write(to_download.content)
-                img = Image.open(my_embed_path)
+            img = Image.open(my_embed_path)
         except:
             print(f"WARNING: Skipping embed file {my_embed_path} due to issues. url: {orig_embed_path}")
         else:
